backend/requirements.txt
# Smart Plant Health Assistant - Python Dependencies
# Production and Development Requirements

# ============================================================================
# Core Framework
# ============================================================================
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Limiter==3.5.0  # Rate limiting
Werkzeug==2.3.7

# ============================================================================
# Database
# ============================================================================
SQLAlchemy==2.0.21
psycopg2-binary==2.9.6  # PostgreSQL adapter

# ============================================================================
# Image Processing & Computer Vision
# ============================================================================
Pillow==10.0.1
scikit-image==0.21.0
opencv-python==4.8.0.76

# ============================================================================
# AI/ML & Deep Learning
# ============================================================================
TensorFlow==2.13.0
Keras==2.13.1
numpy>=1.24.3,<2.0.0
scikit-learn==1.3.1

# ============================================================================
# API & HTTP
# ============================================================================
requests==2.31.0
openai==0.27.8  # OpenAI API client
google-genai  # Google Gemini AI SDK (latest)
google-generativeai  # Google Gemini AI SDK (legacy/stable)

# ============================================================================
# Configuration & Environment
# ============================================================================
python-dotenv==1.0.0

# ============================================================================
# Authentication & Security
# ============================================================================
PyJWT==2.8.0
bcrypt==4.0.1

# ============================================================================
# PDF Generation & Reports
# ============================================================================
reportlab==4.0.5

# ============================================================================
# Background Tasks & Scheduling
# ============================================================================
APScheduler==3.10.4

# ============================================================================
# Email Support
# ============================================================================
Flask-Mail==0.9.1

# ============================================================================
# Push Notifications
# ============================================================================
pywebpush==1.14.0

# ============================================================================
# Data Processing
# ============================================================================
pandas==2.0.3
python-dateutil==2.8.2

# ============================================================================
# Production Server
# ============================================================================
gunicorn==21.2.0
gevent==23.9.1

# ============================================================================
# Development & Testing (optional)
# ============================================================================
# pytest==7.4.0
# pytest-cov==4.1.0
# black==23.7.0
# pylint==2.17.5
# flake8==6.0.0
# pytest-flask==1.2.0

# ============================================================================
# Monitoring & Logging (optional)
# ============================================================================
# python-json-logger==2.0.7
# sentry-sdk==1.28.1

